backend/requirements.txt
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-PyMongo==2.3.0
requests
gunicorn
flask_cors
python-dotenv